@@ -42,7 +42,7 @@
   },
   "devDependencies": {
     "@acme/api": "workspace:^0.1.0",
-    "@acme/eslint-config": "workspace:^0.2.0",
+    "@acme/eslint-config": "workspace:^0.3.0",
     "@acme/prettier-config": "workspace:^0.1.0",
     "@acme/tailwind-config": "workspace:^0.1.0",
     "@acme/tsconfig": "workspace:^0.1.0",
@@ -51,20 +51,10 @@
     "@babel/runtime": "^7.24.5",
     "@types/babel__core": "^7.20.5",
     "@types/react": "^18.3.2",
-    "eslint": "^8.57.0",
+    "eslint": "^9.3.0",
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.5"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "@acme/eslint-config/base",
-      "@acme/eslint-config/react"
-    ],
-    "ignorePatterns": [
-      "expo-plugins/**"
-    ]
   },
   "prettier": "@acme/prettier-config"
 }
