@@ -10,7 +10,7 @@
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
     "format": "prettier --check . --ignore-path ../../.gitignore",
-    "lint": "eslint .",
+    "lint": "eslint",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
@@ -25,18 +25,12 @@
     "next": "^14.2.3"
   },
   "devDependencies": {
-    "@acme/eslint-config": "workspace:^0.2.0",
+    "@acme/eslint-config": "workspace:^0.3.0",
     "@acme/prettier-config": "workspace:^0.1.0",
     "@acme/tsconfig": "workspace:^0.1.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.3.0",
     "prettier": "^3.2.5",
     "typescript": "^5.4.5"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "@acme/eslint-config/base"
-    ]
   },
   "prettier": "@acme/prettier-config"
 }
