@@ -1,10 +1,17 @@
-/** @type {import('eslint').Linter.Config} */
-const config = {
-  extends: ["plugin:@next/next/core-web-vitals"],
-  rules: {
-    "@next/next/no-html-link-for-pages": "off",
-    "@typescript-eslint/require-await": "off",
-  },
-};
+import nextPlugin from "@next/eslint-plugin-next";
 
-module.exports = config;
+/** @type {Awaited<import('typescript-eslint').Config>} */
+export default [
+  {
+    files: ["**/*.ts", "**/*.tsx"],
+    plugins: {
+      "@next/next": nextPlugin,
+    },
+    rules: {
+      ...nextPlugin.configs.recommended.rules,
+      ...nextPlugin.configs["core-web-vitals"].rules,
+      // TypeError: context.getAncestors is not a function
+      "@next/next/no-duplicate-head": "off",
+    },
+  },
+];
