@@ -237,7 +237,7 @@ export function DataTableDemo() {
       <div className="flex items-center justify-between py-4">
         <Input
           placeholder="Filter by name"
-          value={(table.getColumn("name")?.getFilterValue() as string) ?? ""}
+          value={table.getColumn("name")?.getFilterValue() as string}
           onChange={(event) =>
             table.getColumn("name")?.setFilterValue(event.target.value)
           }
