@@ -6,7 +6,7 @@ import { createCaller, createTRPCContext } from "@acme/api";
  * This wraps the `createTRPCContext` helper and provides the required context for the tRPC API when
  * handling a tRPC call from a React Server Component.
  */
-const createContext = cache(async () => {
+const createContext = cache(() => {
   return createTRPCContext();
 });
 
