@@ -1,24 +1,23 @@
-/** @type {import('eslint').Linter.Config} */
-const config = {
-  extends: [
-    "plugin:react/recommended",
-    "plugin:react-hooks/recommended",
-    "plugin:jsx-a11y/recommended",
-  ],
-  rules: {
-    "react/prop-types": "off",
-  },
-  globals: {
-    React: "writable",
-  },
-  settings: {
-    react: {
-      version: "detect",
+import { fixupPluginRules } from "@eslint/compat";
+import reactPlugin from "eslint-plugin-react";
+import hooksPlugin from "eslint-plugin-react-hooks";
+
+/** @type {Awaited<import('typescript-eslint').Config>} */
+export default [
+  {
+    files: ["**/*.ts", "**/*.tsx"],
+    plugins: {
+      react: reactPlugin,
+      "react-hooks": fixupPluginRules(hooksPlugin),
+    },
+    rules: {
+      ...reactPlugin.configs["jsx-runtime"].rules,
+      ...hooksPlugin.configs.recommended.rules,
+    },
+    languageOptions: {
+      globals: {
+        React: "writable",
+      },
     },
   },
-  env: {
-    browser: true,
-  },
-};
-
-module.exports = config;
+];
